major updates
One time registration only for applicants including all details they needed the form - DONE

If the applicant already registered for one program or farm, they cannot register to another one. - ON PROCESS

On Candidates list, it should also appear on what farm the applicant registered. - ON PROCESS

remove status (pending, etc.) - if applicant will register, they will be directly added to Candidates List. - ON PROCESS

If Farm capacity is full, it should show that it is full already and other applicants or new applicants cant register to that program or farm. - ON PROCESS